Changed ArrayType.size from Option[Int] to Option[Long]
- This removes a barrier for arrays with a lot of elements
- Changed constructions and deconstructions of ArrayType to use 64-bit numbers

+++ core/src/main/scala/tungsten/Parser.scala
@@ -426,7 +426,7 @@ object Parser extends Parsers with ImplicitConversions {
   lazy val arrayTy: Parser[ArrayType] = {
     def arraySize = {
       ("?" ^^^ None) |
-      (integer ^^ { case i => Some(i.toInt) })
+      (integer ^^ { case i => Some(i) })
     }
     "[" ~> (arraySize <~ "x") ~ ty <~ "]" ^^ { case s ~ ety => ArrayType(s, ety) }
   }
